feat: Update nodes and links models
Add a new model CommonComment to keep the concurrency comment data.
Use the concurrency comment data to update the nodes and links data which is consumed by frontend.

+++ graph/models.py
@@ -1,5 +1,7 @@
 from django.db import models
 
+from data_api.models import User, Article
+
 # Create your models here.
 class Node(models.Model):
     nodeId = models.CharField(max_length=12, primary_key=True)
@@ -8,4 +10,4 @@ class Node(models.Model):
 class Link(models.Model):
     source = models.ForeignKey(Node, on_delete=models.CASCADE, related_name='out_edge')
     target = models.ForeignKey(Node, on_delete=models.CASCADE, related_name='in_edge')
-    value = models.IntegerField()
+    value = models.FloatField()
